apps/tgbot/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lumen-tgbot"
version = "0.1.0"
description = "Lumen Telegram bot — 菜单驱动生图、任务列表、失败重试"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.13,<4.0",
    "aiohttp-socks>=0.10,<1.0",  # aiogram → SOCKS5（RFC 代理走 api.telegram.org）
    "httpx>=0.27,<1.0",
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "redis>=5.0,<7.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
